feat: add parallel condition checking decorator
